I updated the new IsapiAgent and MapAgentCommon Visual Studio projects to be consistent with our other projects.
Also got rid of warnings in IsapiAgent.


git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@1945 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Web/src/IsapiAgent/IsapiPostParser.cpp
@@ -240,7 +240,7 @@ void IsapiPostParser::PopulateData(char* partHdrEnd, char** curBuf, char* endBuf
         // Note:  dataEnd tag always start with "\r\n--" (see above)
         char* dataStart = partHdrEnd + 4;
         char* dataEnd = dataStart;
-        int dataStartLen = strlen(dataStart);
+        size_t dataStartLen = strlen(dataStart);
         char match0 = dataEndTag[0];
         char match1 = dataEndTag[1];
         char match2 = dataEndTag[2];
